Use ZEO in development mode too.
(we will need it once we start running separated process to run tasks,
like sending queued emails)

Adapted bin/start so it handles starting/stopping ZEO instances.

Added some messages to be shown to the user when running the start script,
helping understanding what the script is doing.

Run pserve in the background, instead of leaving it there on the shell.

Added a bin/stop script that stops pserve and ZEO.